目的地管理添加修改权重的接口

+++ src/main/java/cn/com/wanwei/bic/entity/DestinationEntity.java
@@ -59,7 +59,7 @@ public class DestinationEntity extends BaseEntity{
     private String entertainmentIntroduce;
 
     @ApiModelProperty(value = "权重")
-    private Float weight;
+    private Integer weight;
 
     @ApiModelProperty(value = "组织机构编码")
     private String deptCode;
